Add OTLP translation
This adds the protocol/otlp package that can translate from OTLP metric types
to SignalFx datapoints.  It follows the same conversion process as the OTEL
collector, but that code cannot be reused due to the internal format used within
the OTEL Collector.

This also adds a decoder that reads from an HTTP request and sends to a
sink.